requirements.txt
opencv-python
torchvision
matplotlib
IPython
requests
timm
bitstring
jellyfish
humanize
psutil